+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -231,7 +231,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K9HC1ZDXDQ3AW0GWR1R9FTJZ"),
+  _id: z.string().default("01K9P3D694Z8HCEFYBBQWZWV84"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -274,7 +274,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01K9HC1ZDXDQ3AW0GWR1R9FTJZ"),
+  id: z.string().default("01K9P3D694Z8HCEFYBBQWZWV84"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,