+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -145,7 +145,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K9HC1ZDYKYTNPDKXX1X8KK9P"),
+  _id: z.string().default("01K9P3D696C8W5VCDZSC9AB5DK"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -188,7 +188,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K9HC1ZDYKYTNPDKXX1X8KK9P"),
+  id: z.string().default("01K9P3D696C8W5VCDZSC9AB5DK"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,