+++ packages/orq-rc/RELEASES.md
@@ -4869,3 +4869,13 @@ Based on:
 - [typescript v4.0.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.17 - packages/orq-rc
+
+## 2025-11-10 05:23:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.651.0 (2.745.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.18 - packages/orq-rc